==> code/classification/02_rewrite_food_tag_files.ipynb <==
import os
import sys
import pandas as pd

# 添加项目根目录到 Python 路径
project_root = '/date0/crwu/zjts_foodname'
if project_root not in sys.path:
    sys.path.insert(0, project_root)

if project_root not in sys.path:
    sys.path.insert(0, project_root)

from packages.file_deal import read_dataset

food_tags_path = r'/date0/crwu/zjts_foodname/data/classification/data/food_tags.csv'
data = {
    '主要食材种类': [
'米类','面粉或小麦类','玉米类','杂粮类','叶菜类','根茎类','瓜果类','茄果类','菌菇类','大豆或豆类','高纤维豆类','高淀粉豆类','薯类','豆腐或豆干类','豆浆或腐竹类','发酵豆制品','鸡蛋','鸭蛋','鹌鹑蛋','其他禽蛋','猪肉','牛肉','羊肉','其他畜肉','鸡肉','鸭肉','鹅肉','其他禽肉','畜肉内脏','禽类内脏','淡水鱼','海水鱼','虾类','蟹类','贝类','软体类','爬行或两栖类','海产干货','仁果或核果类','浆果类','柑橘类','热带或亚热带水果','瓜类','奶或奶油类','奶酪或芝士类','发酵乳品','坚果或种子类','茶或咖啡或可可类','藻类','鹅蛋','葱蒜或洋葱类','其他','不适用',
        ],
    '烹饪方式': [
'炒','爆','烧','焖','卤','烩','蒸','煮','炖','涮','炸','溜','凉拌','烤','煎','熏','焗','腌','醉','焯','汆','其他',
        ],
    '菜系（地域）分类': [
'川菜','粤菜','鲁菜','闽菜','苏菜','浙菜','湘菜','徽菜','京菜','本帮菜','东北菜','客家菜','西北菜','云南菜','港式/茶餐厅','台式','家常/不区分菜系','西餐','日料','韩料','东南亚菜',
        ],
    '消费场景': [
'家常菜','餐馆菜','宴席菜/大菜','小吃/街头美食','预制菜/工业化食品','其他',
        ],
    '营养建议分类': [
'主食','淀粉类','高糖','高油','无',
        ],
    '记忆/推理菜':[
'推理菜','记忆菜',
        ],
    '早中晚餐消费场景':[
'早餐','午餐或晚餐','夜宵','下午茶/零食','其他',
        ],
    '食用频率':[
'1','2','3','4','5',
        ],
    '熟悉程度':[
'1','2','3','4','5',
        ],
    }
max_len = 0
for key,vals in data.items():
    max_len = max (max_len, len(vals))
print(max_len)

for key,vals in data.items():
    if len(vals) < max_len:
        vals.extend([pd.NA] * (max_len - len(vals)))
df_tags = pd.DataFrame(data)

# df_tags.to_csv(food_tags_path, index=False,encoding='gbk')

food_depth_path = r'/date0/crwu/zjts_foodname/data/classification/data/food_depth.csv'
depth_df = read_dataset(food_depth_path)

depth_md = '''| 第一级别 | 第二级别 | 第三级别 | 第四级别 |
| :--- | :--- | :--- | :--- |'''
last_level_1 = ''
last_level_2 = ''
last_level_3 = ''
last_level_4 = ''
for row in depth_df.iterrows():
    level_1, level_2, level_3,level_4 = row[1]
    # print(level_1, level_2, level_3,level_4)
    if level_1 == last_level_1:
        current_level_1 = ' '
    else:
        current_level_1 = f' {level_1} '
        last_level_1 = level_1
    if level_2 == last_level_2:
        current_level_2 = ' '
    else:
        current_level_2 = f' {level_2} '
        last_level_2 = level_2
    if level_3 == last_level_3:
        current_level_3 = ' '
    else:
        current_level_3 = f' {level_3} '
        last_level_3 = level_3
    if level_4 == last_level_4:
        current_level_4 = ' '
    else:
        current_level_4 = f' {level_4} '
        last_level_4 = level_4

    depth_md += f'\n|{current_level_1}|{current_level_2}|{current_level_3}|{current_level_4}|'
print(depth_md)
        

food_tags_path = r'/date0/crwu/zjts_foodname/data/classification/data/food_tags.csv'
tags_df = read_dataset(food_tags_path)

for column in tags_df.columns:
    print(f'[{column}]')
    # print(tags_df[column])
    print(','.join(list(([str(tag).strip('.0') for tag in tags_df[column] if not pd.isna(tag)]))))
    print()
